+++ packages/bridge-controller/src/index.ts
@@ -100,7 +100,7 @@ export {
   isSolanaChainId,
   getNativeAssetForChainId,
   getDefaultBridgeControllerState,
-  isCrossChainTx
+  isCrossChain
 } from './utils/bridge';
 
 export { isValidQuoteRequest, formatEtaInMinutes } from './utils/quote';